M2/Ecommerce/requirements.txt
Flask==3.1.1            
SQLAlchemy==2.0.41      
psycopg2-binary==2.9.10 
redis==6.4.0             
PyJWT==2.10.1           
python-dotenv==1.1.1     
bcrypt==4.3.0            
pytest==8.4.1